[Gold II] Title: 선분 교차 2, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/17387. 선분 교차 2/선분 교차 2.cc
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long ccw(pair<long, long> a, pair<long, long> b, pair<long, long> c)
+int ccw(pair<long, long> a,pair<long, long> b,pair<long, long> c)
 {
     long long D;
     D = a.first * b.second + b.first * c.second + c.first * a.second - a.second * b.first - b.second * c.first - c.second * a.first;
@@ -20,12 +20,12 @@ int main()
     cin.tie(NULL);
     cout.tie(NULL);
 
-    pair<int, int> a, b, c, d;
+    pair<long, long> a, b, c, d;
     cin >> a.first >> a.second >> b.first >> b.second;
     cin >> c.first >> c.second >> d.first >> d.second;
 
-    long long D1 = ccw(a, b, c) * ccw(a, b, d);
-    long long D2 = ccw(c, d, a) * ccw(c, d, b);
+    int D1 = ccw(a, b, c) * ccw(a, b, d);
+    int D2 = ccw(c, d, a) * ccw(c, d, b);
 
     if (D1 == 0 && D2 == 0)
     {
